+++ corehq/apps/app_manager/schemas/document/form_action.py
@@ -1,0 +1,129 @@
+from __future__ import absolute_import
+from __future__ import unicode_literals
+
+from dimagi.ext.couchdbkit import (
+    BooleanProperty,
+    DateTimeProperty,
+    DecimalProperty,
+    DictProperty,
+    Document,
+    DocumentSchema,
+    FloatProperty,
+    IntegerProperty,
+    ListProperty,
+    SchemaDictProperty,
+    SchemaListProperty,
+    SchemaProperty,
+    StringListProperty,
+    StringProperty,
+)
+
+
+class FormActionCondition(DocumentSchema):
+    """
+    The condition under which to open/update/close a case/referral
+
+    Either {'type': 'if', 'question': '/xpath/to/node', 'answer': 'value'}
+    in which case the action takes place if question has answer answer,
+    or {'type': 'always'} in which case the action always takes place.
+    """
+    type = StringProperty(choices=["if", "always", "never"], default="never")
+    question = StringProperty()
+    answer = StringProperty()
+    operator = StringProperty(choices=['=', 'selected', 'boolean_true'], default='=')
+
+    def is_active(self):
+        return self.type in ('if', 'always')
+
+
+class FormAction(DocumentSchema):
+    """
+    Corresponds to Case XML
+
+    """
+    condition = SchemaProperty(FormActionCondition)
+
+    def is_active(self):
+        return self.condition.is_active()
+
+    @classmethod
+    def get_action_paths(cls, action):
+        if action.condition.type == 'if':
+            yield action.condition.question
+
+        for __, path in cls.get_action_properties(action):
+            yield path
+
+    @classmethod
+    def get_action_properties(self, action):
+        action_properties = action.properties()
+        if 'name_path' in action_properties and action.name_path:
+            yield 'name', action.name_path
+        if 'case_name' in action_properties:
+            yield 'name', action.case_name
+        if 'external_id' in action_properties and action.external_id:
+            yield 'external_id', action.external_id
+        if 'update' in action_properties:
+            for name, path in action.update.items():
+                yield name, path
+        if 'case_properties' in action_properties:
+            for name, path in action.case_properties.items():
+                yield name, path
+        if 'preload' in action_properties:
+            for path, name in action.preload.items():
+                yield name, path
+
+
+class UpdateCaseAction(FormAction):
+
+    update = DictProperty()
+
+
+class PreloadAction(FormAction):
+
+    preload = DictProperty()
+
+    def is_active(self):
+        return bool(self.preload)
+
+
+class UpdateReferralAction(FormAction):
+
+    followup_date = StringProperty()
+
+    def get_followup_date(self):
+        if self.followup_date:
+            return "if(date({followup_date}) >= date(today()), {followup_date}, date(today() + 2))".format(
+                followup_date=self.followup_date,
+            )
+        return self.followup_date or "date(today() + 2)"
+
+
+class OpenReferralAction(UpdateReferralAction):
+
+    name_path = StringProperty()
+
+
+class OpenCaseAction(FormAction):
+
+    name_path = StringProperty()
+    external_id = StringProperty()
+
+
+class OpenSubCaseAction(FormAction, IndexedSchema):
+
+    case_type = StringProperty()
+    case_name = StringProperty()
+    reference_id = StringProperty()
+    case_properties = DictProperty()
+    repeat_context = StringProperty()
+    # relationship = "child" for index to a parent case (default)
+    # relationship = "extension" for index to a host case
+    relationship = StringProperty(choices=['child', 'extension'], default='child')
+
+    close_condition = SchemaProperty(FormActionCondition)
+
+    @property
+    def form_element_name(self):
+        return 'subcase_{}'.format(self.id)
+
